api/requirements.txt
blinker==1.9.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
# Flask==3.1.0
fonttools==4.56.0
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
# matplotlib==3.10.1
networkx==3.4.2
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
tifffile==2025.2.18
Unidecode==1.3.8
Werkzeug==3.1.3
maturin==1.8.3
underthesea==6.8.4
pandas==2.2.3
openpyxl==3.1.5
black==25.1.0
isort==6.0.1
uvicorn==0.34.0
fastapi==0.115.12
httpx==0.28.1
PyJWT==2.10.1
# rich==13.9.4
# transformers==4.49.0
# torch==2.6.0
# torchvision==0.21.0